refactor(dtypes): make numpy and pandas optional dependencies
